Rename LA64Base -> LoongArchBase, LA_LSX -> Lsx, LA_LASX -> Lasx. Update the API within the LoongArchBase class.
