+++ src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/LoongArch/Lasx.PlatformNotSupported.cs
