+++ src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/LoongArch/Lsx.PlatformNotSupported.cs
